+++ test/sanity/validate-modules/ignore.txt
@@ -486,27 +486,6 @@ lib/ansible/modules/cloud/vmware/vca_vapp.py E322
 lib/ansible/modules/cloud/vmware/vca_vapp.py E324
 lib/ansible/modules/cloud/vmware/vmware_guest.py E322
 lib/ansible/modules/cloud/vmware/vsphere_copy.py E322
-lib/ansible/modules/cloud/vultr/vr_account_facts.py E324
-lib/ansible/modules/cloud/vultr/vr_dns_domain.py E323
-lib/ansible/modules/cloud/vultr/vr_dns_domain.py E324
-lib/ansible/modules/cloud/vultr/vr_dns_record.py E324
-lib/ansible/modules/cloud/vultr/vr_dns_record.py E326
-lib/ansible/modules/cloud/vultr/vr_firewall_group.py E324
-lib/ansible/modules/cloud/vultr/vr_firewall_rule.py E323
-lib/ansible/modules/cloud/vultr/vr_firewall_rule.py E324
-lib/ansible/modules/cloud/vultr/vr_server.py E324
-lib/ansible/modules/cloud/vultr/vr_server.py E325
-lib/ansible/modules/cloud/vultr/vr_ssh_key.py E324
-lib/ansible/modules/cloud/vultr/vr_startup_script.py E324
-lib/ansible/modules/cloud/vultr/vr_user.py E322
-lib/ansible/modules/cloud/vultr/vr_user.py E324
-lib/ansible/modules/cloud/vultr/vr_user.py E325
-lib/ansible/modules/cloud/vultr/vr_user.py E326
-lib/ansible/modules/cloud/webfaction/webfaction_app.py E321
-lib/ansible/modules/cloud/webfaction/webfaction_db.py E321
-lib/ansible/modules/cloud/webfaction/webfaction_domain.py E321
-lib/ansible/modules/cloud/webfaction/webfaction_mailbox.py E321
-lib/ansible/modules/cloud/webfaction/webfaction_site.py E321
 lib/ansible/modules/clustering/consul.py E322
 lib/ansible/modules/clustering/consul.py E324
 lib/ansible/modules/clustering/consul_acl.py E325
